Don't run down if not up

+++ scripts/stop.sh
@@ -11,7 +11,9 @@ if [ "$1" = "--help" ]; then
   exit
 fi
 
-GROUP_ID=$(id -g) USER_ID=$(id -u) docker-compose down
+if [ ! -z "$(docker network ls | awk '/formal-systems/ {print $2}')" ]; then
+  GROUP_ID=$(id -g) USER_ID=$(id -u) docker-compose down
+fi
 
 if [ "$1" == "--clean" ] && [ "$(ls -A database-files)" ]; then
   rm database-files/.[!.]* database-files/* -r
